Admin - Extract MotoAPI functionality to dedicated module
