+++ moto/moto_api/_internal/__init__.py
@@ -1,0 +1,4 @@
+from .models import moto_api_backend
+
+
+moto_api_backends = {"global": moto_api_backend}
