+++ moto/moto_api/_internal/responses.py
@@ -1,0 +1,67 @@
+import json
+
+from moto import settings
+from moto.core.responses import ActionAuthenticatorMixin, BaseResponse
+
+
+class MotoAPIResponse(BaseResponse):
+    def reset_response(
+        self, request, full_url, headers
+    ):  # pylint: disable=unused-argument
+        if request.method == "POST":
+            from .models import moto_api_backend
+
+            moto_api_backend.reset()
+            return 200, {}, json.dumps({"status": "ok"})
+        return 400, {}, json.dumps({"Error": "Need to POST to reset Moto"})
+
+    def reset_auth_response(
+        self, request, full_url, headers
+    ):  # pylint: disable=unused-argument
+        if request.method == "POST":
+            previous_initial_no_auth_action_count = (
+                settings.INITIAL_NO_AUTH_ACTION_COUNT
+            )
+            settings.INITIAL_NO_AUTH_ACTION_COUNT = float(request.data.decode())
+            ActionAuthenticatorMixin.request_count = 0
+            return (
+                200,
+                {},
+                json.dumps(
+                    {
+                        "status": "ok",
+                        "PREVIOUS_INITIAL_NO_AUTH_ACTION_COUNT": str(
+                            previous_initial_no_auth_action_count
+                        ),
+                    }
+                ),
+            )
+        return 400, {}, json.dumps({"Error": "Need to POST to reset Moto Auth"})
+
+    def model_data(self, request, full_url, headers):  # pylint: disable=unused-argument
+        from moto.core.models import model_data
+
+        results = {}
+        for service in sorted(model_data):
+            models = model_data[service]
+            results[service] = {}
+            for name in sorted(models):
+                model = models[name]
+                results[service][name] = []
+                for instance in model.instances:
+                    inst_result = {}
+                    for attr in dir(instance):
+                        if not attr.startswith("_"):
+                            try:
+                                json.dumps(getattr(instance, attr))
+                            except TypeError:
+                                pass
+                            else:
+                                inst_result[attr] = getattr(instance, attr)
+                    results[service][name].append(inst_result)
+        return 200, {"Content-Type": "application/javascript"}, json.dumps(results)
+
+    def dashboard(self, request, full_url, headers):  # pylint: disable=unused-argument
+        from flask import render_template
+
+        return render_template("dashboard.html")
